requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
solana>=0.30.0
solders>=0.18.0
base58>=2.1.0

